feat(pagination): new default and Pager story
