+++ packages/pagination/index.js
@@ -1,7 +1,6 @@
-import AsyncPagination from './AsyncPagination';
 import Pagination from './Pagination';
-import PaginationItem from './PaginationItem';
+
+export * from './Pager';
+export AsyncPagination from './AsyncPagination';
 
 export default Pagination;
-
-export { AsyncPagination, PaginationItem };
